AI.Interview/requirements.txt
streamlit
openai
PyMuPDF
opencv-python
docx2txt
Pillow
numpy
# pip install -r requirements.txt
# streamlit run app.py
